[ YC-632, YC-734 ] YUM webapp, Improve access to carriers for shop owners
- improved error notification dialog
- added carrier and SLA section
- some improvements in vo to dto conversion (added converters)

+++ manager/jam-services/src/main/java/org/yes/cart/service/vo/impl/VoLocationServiceImpl.java
@@ -16,14 +16,15 @@
 
 package org.yes.cart.service.vo.impl;
 
-import com.inspiresoftware.lib.dto.geda.assembler.Assembler;
 import com.inspiresoftware.lib.dto.geda.assembler.DTOAssembler;
+import org.springframework.security.access.AccessDeniedException;
 import org.yes.cart.domain.dto.CountryDTO;
 import org.yes.cart.domain.dto.StateDTO;
 import org.yes.cart.domain.vo.VoCountry;
 import org.yes.cart.domain.vo.VoState;
 import org.yes.cart.service.dto.DtoCountryService;
 import org.yes.cart.service.dto.DtoStateService;
+import org.yes.cart.service.federation.FederationFacade;
 import org.yes.cart.service.vo.VoLocationService;
 
 import java.util.ArrayList;
@@ -38,15 +39,15 @@ public class VoLocationServiceImpl implements VoLocationService {
 
     private final DtoCountryService dtoCountryService;
     private final DtoStateService dtoStateService;
-    private final Assembler simpleVoCountryAssembler;
-    private final Assembler simpleVoStateAssembler;
+
+    private final FederationFacade federationFacade;
 
     public VoLocationServiceImpl(final DtoCountryService dtoCountryService,
-                                 final DtoStateService dtoStateService) {
+                                 final DtoStateService dtoStateService,
+                                 final FederationFacade federationFacade) {
         this.dtoCountryService = dtoCountryService;
         this.dtoStateService = dtoStateService;
-        this.simpleVoCountryAssembler = DTOAssembler.newAssembler(VoCountry.class, CountryDTO.class);
-        this.simpleVoStateAssembler = DTOAssembler.newAssembler(VoState.class, StateDTO.class);
+        this.federationFacade = federationFacade;
     }
 
     /** {@inheritDoc} */
@@ -54,7 +55,7 @@ public class VoLocationServiceImpl implements VoLocationService {
     public List<VoCountry> getAllCountries() throws Exception {
         final List<CountryDTO> all = dtoCountryService.getAll();
         final List<VoCountry> rez = new ArrayList<>(all.size());
-        simpleVoCountryAssembler.assembleDtos(rez, all, null, null);
+        DTOAssembler.newAssembler(VoCountry.class, CountryDTO.class).assembleDtos(rez, all, null, null);
         return rez;
     }
 
@@ -64,7 +65,7 @@ public class VoLocationServiceImpl implements VoLocationService {
         final CountryDTO countryDTO = dtoCountryService.getById(id);
         if (countryDTO != null) {
             final VoCountry voCountry = new VoCountry();
-            simpleVoCountryAssembler.assembleDto(voCountry, countryDTO, null, null);
+            DTOAssembler.newAssembler(VoCountry.class, CountryDTO.class).assembleDto(voCountry, countryDTO, null, null);
             return voCountry;
         }
         return null;
@@ -72,29 +73,38 @@ public class VoLocationServiceImpl implements VoLocationService {
 
     /** {@inheritDoc} */
     @Override
-    public VoCountry createCountry(final VoCountry voCategory) throws Exception {
-        CountryDTO countryDTO = dtoCountryService.getNew();
-        simpleVoCountryAssembler.assembleEntity(voCategory, countryDTO, null, null);
-        countryDTO = dtoCountryService.create(countryDTO);
-        return getCountryById(countryDTO.getCountryId());
+    public VoCountry createCountry(final VoCountry vo) throws Exception {
+        if (federationFacade.isCurrentUserSystemAdmin()) {
+            CountryDTO countryDTO = dtoCountryService.getNew();
+            DTOAssembler.newAssembler(VoCountry.class, CountryDTO.class).assembleEntity(vo, countryDTO, null, null);
+            countryDTO = dtoCountryService.create(countryDTO);
+            return getCountryById(countryDTO.getCountryId());
+        } else {
+            throw new AccessDeniedException("Access is denied");
+        }
     }
 
     /** {@inheritDoc} */
     @Override
-    public VoCountry updateCountry(final VoCountry voCategory) throws Exception {
-        CountryDTO countryDTO = dtoCountryService.getById(voCategory.getCountryId());
-        if (countryDTO != null) {
-            simpleVoCountryAssembler.assembleEntity(voCategory, countryDTO, null, null);
+    public VoCountry updateCountry(final VoCountry vo) throws Exception {
+        CountryDTO countryDTO = dtoCountryService.getById(vo.getCountryId());
+        if (countryDTO != null && federationFacade.isCurrentUserSystemAdmin()) {
+            DTOAssembler.newAssembler(VoCountry.class, CountryDTO.class).assembleEntity(vo, countryDTO, null, null);
             countryDTO = dtoCountryService.update(countryDTO);
             return getCountryById(countryDTO.getCountryId());
+        } else {
+            throw new AccessDeniedException("Access is denied");
         }
-        return null;
     }
 
     /** {@inheritDoc} */
     @Override
     public void removeCountry(final long id) throws Exception {
-        dtoCountryService.remove(id);
+        if (federationFacade.isCurrentUserSystemAdmin()) {
+            dtoCountryService.remove(id);
+        } else {
+            throw new AccessDeniedException("Access is denied");
+        }
     }
 
     /** {@inheritDoc} */
@@ -104,7 +114,7 @@ public class VoLocationServiceImpl implements VoLocationService {
         final List<VoState> rez = new ArrayList<>();
         if (countryDTO != null) {
             final List<StateDTO> all = dtoStateService.findByCountry(countryDTO.getCountryCode());
-            simpleVoStateAssembler.assembleDtos(rez, all, null, null);
+            DTOAssembler.newAssembler(VoState.class, StateDTO.class).assembleDtos(rez, all, null, null);
         }
         return rez;
     }
@@ -115,7 +125,7 @@ public class VoLocationServiceImpl implements VoLocationService {
         final StateDTO stateDTO = dtoStateService.getById(id);
         if (stateDTO != null) {
             final VoState voState = new VoState();
-            simpleVoStateAssembler.assembleDto(voState, stateDTO, null, null);
+            DTOAssembler.newAssembler(VoState.class, StateDTO.class).assembleDto(voState, stateDTO, null, null);
             return voState;
         }
         return null;
@@ -123,29 +133,38 @@ public class VoLocationServiceImpl implements VoLocationService {
 
     /** {@inheritDoc} */
     @Override
-    public VoState createState(final VoState voState) throws Exception {
-        StateDTO stateDTO = dtoStateService.getNew();
-        simpleVoStateAssembler.assembleEntity(voState, stateDTO, null, null);
-        stateDTO = dtoStateService.create(stateDTO);
-        return getStateById(stateDTO.getStateId());
+    public VoState createState(final VoState vo) throws Exception {
+        if (federationFacade.isCurrentUserSystemAdmin()) {
+            StateDTO stateDTO = dtoStateService.getNew();
+            DTOAssembler.newAssembler(VoState.class, StateDTO.class).assembleEntity(vo, stateDTO, null, null);
+            stateDTO = dtoStateService.create(stateDTO);
+            return getStateById(stateDTO.getStateId());
+        } else {
+            throw new AccessDeniedException("Access is denied");
+        }
     }
 
     /** {@inheritDoc} */
     @Override
-    public VoState updateState(final VoState voState) throws Exception {
-        StateDTO stateDTO = dtoStateService.getById(voState.getStateId());
-        if (stateDTO != null) {
-            simpleVoStateAssembler.assembleEntity(voState, stateDTO, null, null);
+    public VoState updateState(final VoState vo) throws Exception {
+        StateDTO stateDTO = dtoStateService.getById(vo.getStateId());
+        if (stateDTO != null && federationFacade.isCurrentUserSystemAdmin()) {
+            DTOAssembler.newAssembler(VoState.class, StateDTO.class).assembleEntity(vo, stateDTO, null, null);
             stateDTO = dtoStateService.update(stateDTO);
             return getStateById(stateDTO.getStateId());
+        } else {
+            throw new AccessDeniedException("Access is denied");
         }
-        return null;
     }
 
     /** {@inheritDoc} */
     @Override
     public void removeState(final long id) throws Exception {
-        dtoStateService.remove(id);
+        if (federationFacade.isCurrentUserSystemAdmin()) {
+            dtoStateService.remove(id);
+        } else {
+            throw new AccessDeniedException("Access is denied");
+        }
     }
 
 }
